Extract GTModelTexture test method to class
